perf(database): Upgrade id to uint64

+++ internal/database/db.go
@@ -11,13 +11,13 @@ type DBRepository interface {
 	Ping(ctx context.Context) error
 	UserCreate(ctx context.Context, user models.User) (*models.User, error)
 	UserGetAll(ctx context.Context) ([]*models.User, error)
-	UserGetByID(ctx context.Context, id int) (*models.User, error)
-	UserDeleteByID(ctx context.Context, id int) error
+	UserGetByID(ctx context.Context, id uint64) (*models.User, error)
+	UserDeleteByID(ctx context.Context, id uint64) error
 	UserUpdate(ctx context.Context, user models.UserUpdate) (*models.User, error)
 
 	PostCreate(ctx context.Context, post models.Post) (*models.Post, error)
 	PostGetAll(ctx context.Context) ([]*models.Post, error)
-	PostGetByID(ctx context.Context, id int) (*models.Post, error)
-	PostDeleteByID(ctx context.Context, id int) error
+	PostGetByID(ctx context.Context, id uint64) (*models.Post, error)
+	PostDeleteByID(ctx context.Context, id uint64) error
 	PostUpdate(ctx context.Context, post models.PostUpdate) (*models.Post, error)
 }
